New levels, sprite-changes and more!
Only 1 level needs to be made to finish Icy Island now.
Yeti is fixed and his boss fight is now pretty cool but pretty quick, but that makes sense for the first boss, it's the same in New Super Mario Bros Wii's Koopaling fights or New Super Mario Bros U's Boom Boom sprites where it's really easy and really quick if you know what you're doing, but since this add-on is aimed at those who are new to SuperTux, the players probably won't know what they're doing so this Yeti fight would be a pretty fun boss fight for them that doesn't last as long as the normal SuperTux Yeti fight.

Levels are also fixed a bit more, also, trees are recolored and some snow for them is in the decals now.

Some songs are also now in different places because why not?